app/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile app/requirements.in
#
attrs==19.3.0             # via pytest
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
django==2.2.7
djangorestframework==3.10.3
idna==2.8                 # via requests
importlib-metadata==1.2.0  # via pluggy, pytest
more-itertools==8.0.2     # via pytest, zipp
packaging==19.2           # via pytest
pathlib2==2.3.5           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.5          # via packaging
pytest==5.3.1
python-decouple==3.3
pytz==2019.3              # via django
requests==2.22.0
six==1.13.0               # via packaging, pathlib2
sqlparse==0.3.0           # via django
urllib3==1.25.7           # via requests
uwsgi==2.0.18
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
